backend/requirements.txt
django>=4.2,<5.0
djangorestframework
django-cors-headers
psycopg2-binary
pymongo
python-dotenv
requests
cloudinary
redis
whitenoise
gunicorn
sentry-sdk
drf-yasg
dj-database-url
django-filter
djangorestframework-simplejwt
google-auth

